% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enable.r
\name{RdymEnable}
\alias{RdymDisable}
\alias{RdymEnable}
\title{RdymEnable}
\usage{
RdymEnable()

RdymDisable()
}
\description{
Enable "did you mean?" error checking.

Disable "did you mean?" error checking.
}
\examples{
\dontrun{
RdymEnable()
nonsense # or something else that doesn't exist and causes an error

RdymDisable()
}
}

